third_party/blink/web_tests/external/wpt/editing/run/insertparagraph_1001-2000-expected.txt

This is a testharness.js-based test.
Found 39 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<script>foo[]bar</script>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><script>foo</script><br></div><div><script>bar</script>baz</div>" but got "<div><br></div><script>foobar</script>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<script>foo[]bar</script>baz" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<script>foo[]bar</script>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><script>foo</script><br></p><p><script>bar</script>baz</p>" but got "<p><br></p><script>foobar</script>baz"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<div style=display:none>foo[]bar</div>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:none\\">foo<br></div><div style=\\"display:none\\">bar<br></div>baz" but got "<div><br></div><div style=\\"display:none\\">foobar</div>baz"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<div style=display:none>foo[]bar</div>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<div style=display:none>foo[]bar</div>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:none\\">foo<br></div><div style=\\"display:none\\">bar<br></div>baz" but got "<div><br></div><div style=\\"display:none\\">foobar</div>baz"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<div style=display:none>foo[]bar</div>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:none\\">foo<br></div><div style=\\"display:none\\">bar<br></div>baz" but got "<p><br></p><div style=\\"display:none\\">foobar</div>baz"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<div style=display:none>foo[]bar</div>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:none\\">foo<br></div><div style=\\"display:none\\">bar<br></div>baz" but got "<p><br></p><div style=\\"display:none\\">foobar</div>baz"
[FAIL] [["insertparagraph",""]] "<listing>foo[]bar</listing>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<listing>foo<br>bar</listing>" but got "<listing>foo</listing><listing>bar</listing>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<dl><dt>{}<br></dt></dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div>" but got "<dl><dt><br></dt><dt><br></dt></dl>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<dl><dt>{}<br></dt></dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><br></p>" but got "<dl><dt><br></dt><dt><br></dt></dl>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<dl><dt>foo<dd>{}<br></dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt></dl><div><br></div>" but got "<dl><dt>foo</dt><dd><br></dd><dd><br></dd></dl>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<dl><dt>foo<dd>{}<br></dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt></dl><p><br></p>" but got "<dl><dt>foo</dt><dd><br></dd><dd><br></dd></dl>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<dl><dt>{}<br><dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div><dl><dd>bar</dd></dl>" but got "<dl><dt><br></dt><dt><br></dt><dd>bar</dd></dl>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<dl><dt>{}<br><dd>bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><br></p><dl><dd>bar</dd></dl>" but got "<dl><dt><br></dt><dt><br></dt><dd>bar</dd></dl>"
[FAIL] [["insertparagraph",""]] "<dl><dt>foo<dd>bar<dl><dt>{}<br><dd>baz</dl></dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>bar</dd><dt><br></dt><dd><dl><dd>baz</dd></dl></dd></dl>" but got "<dl><dt>foo</dt><dd>bar<dl><dt><br></dt><dt><br></dt><dd>baz</dd></dl></dd></dl>"
[FAIL] [["insertparagraph",""]] "<dl><dt>foo<dd>bar<dl><dt>baz<dd>{}<br></dl></dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>bar<dl><dt>baz</dt></dl></dd><dd><br></dd></dl>" but got "<dl><dt>foo</dt><dd>bar<dl><dt>baz</dt><dd><br></dd><dd><br></dd></dl></dd></dl>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<p>foo</p>{}<br>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><div><br></div><div><br></div>" but got "<p>foo</p><br><div><br></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<p>foo</p>{}<br>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p><p><br></p>" but got "<p>foo</p><br><p><br></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "{}<br><p>foo</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div><div><br></div><p>foo</p>" but got "<div><br></div><br><p>foo</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "{}<br><p>foo</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><br></p><p><br></p><p>foo</p>" but got "<p><br></p><br><p>foo</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<p>foo</p>{}<br><h1>bar</h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><div><br></div><div><br></div><h1>bar</h1>" but got "<p>foo</p><div><br></div><br><h1>bar</h1>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<p>foo</p>{}<br><h1>bar</h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p><p><br></p><h1>bar</h1>" but got "<p>foo</p><p><br></p><br><h1>bar</h1>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<h1>foo</h1>{}<br><p>bar</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><div><br></div><div><br></div><p>bar</p>" but got "<h1>foo</h1><div><br></div><br><p>bar</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<h1>foo</h1>{}<br><p>bar</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><p><br></p><p><br></p><p>bar</p>" but got "<h1>foo</h1><p><br></p><br><p>bar</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<h1>foo</h1>{}<br><h2>bar</h2>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><div><br></div><div><br></div><h2>bar</h2>" but got "<h1>foo</h1><div><br></div><br><h2>bar</h2>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<h1>foo</h1>{}<br><h2>bar</h2>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><p><br></p><p><br></p><h2>bar</h2>" but got "<h1>foo</h1><p><br></p><br><h2>bar</h2>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<p>foo</p>{<h1>bar</h1>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><div><br></div><p>baz</p>" but got "<p>foo</p><h1><br></h1><h1>baz</h1>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<p>foo</p>{<h1>bar</h1>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><br></p><p>baz</p>" but got "<p>foo</p><h1><br></h1><h1>baz</h1>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<table><tr><td>foo[]bar</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><div>foo</div><div>bar</div></td></tr></tbody></table>" but got "<table><tbody><tr><td>foo<br>bar</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<table><tr><td>foo[]bar</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><p>foo</p><p>bar</p></td></tr></tbody></table>" but got "<table><tbody><tr><td>foo<br>bar</td></tr></tbody></table>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<blockquote>[]foo</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><div><br></div><div>foo</div></blockquote>" but got "<blockquote><br></blockquote><blockquote>foo</blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<blockquote>[]foo</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><p><br></p><p>foo</p></blockquote>" but got "<blockquote><br></blockquote><blockquote>foo</blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<blockquote>foo[]</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><div>foo</div><div><br></div></blockquote>" but got "<blockquote>foo</blockquote><blockquote><br></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<blockquote>foo[]</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><p>foo</p><p><br></p></blockquote>" but got "<blockquote>foo</blockquote><blockquote><br></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<blockquote>foo[]<br></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><div>foo</div><div><br></div></blockquote>" but got "<blockquote>foo<br></blockquote><blockquote><br></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<blockquote>foo[]<br></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><p>foo</p><p><br></p></blockquote>" but got "<blockquote>foo<br></blockquote><blockquote><br></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<blockquote>foo[]bar</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><div>foo</div><div>bar</div></blockquote>" but got "<blockquote>foo</blockquote><blockquote>bar</blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<blockquote>foo[]bar</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><p>foo</p><p>bar</p></blockquote>" but got "<blockquote>foo</blockquote><blockquote>bar</blockquote>"
Harness: the test ran to completion.
